specialisation/leaffliction/pyproject.toml
[tool.poetry]
name = "leafliction"
version = "0.1.0"
description = ""
authors = ["jreivilo <j.reivilo@outlook.com>"]

[tool.poetry.dependencies]
python = "3.10"
matplotlib = "^3.8.2"
mne = "^1.6.0"
pillow = "^10.2.0"
pandas = "^2.1.4"
numpy = "^1.26.3"
torch = "^2.0.0"
torchvision = "^0.15.0"
plantcv = "^4.0.1"
ipython = "^8.20.0"
tensorflow = "^2.15.0"  # Regular TensorFlow dependency


[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
